fix(gen): nullable and optional arrays JSON encoders generation
